refactoring repositories

+++ sharedRedream-apis/app/Repositories/Impl/WalletRepositoryInterface.php
@@ -7,7 +7,7 @@ interface WalletRepositoryInterface
     public function findById(int $id);
     public function findByUserId(int $user_id);
     public function create(array $data);
-    public function deposit(int $user_id, int $value);
-    public function withdrawal(int $user_id, int $value);
+    public function deposit(int $user_id, float $value);
+    public function withdrawal(int $user_id, float $value);
 
 }
